GetItem implementation

+++ wdm-microservices-project/stocks/src/main/java/wdm/project/repository/StocksRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import wdm.project.dto.Item;
 
 @Repository
-public interface StocksRepository extends JpaRepository<Item, Integer> {
+public interface StocksRepository extends JpaRepository<Item, Long> {
 }
